enabled anti slap

+++ modules/special_case.js
@@ -10,18 +10,18 @@ module.exports = (global_params,global_settings,message_type) => {
 		yush id: 109385289813798912
 		Arciel id: 151255303852261376
 	*/
-/*	if((message_content.includes("slap") || message_content.includes("punch")) && (message_content.includes("kirimaru")) || message_content.includes("<@"+global_params.env_vars.OWNER_ID+">")){
+	if((message_content.includes("slap") || message_content.includes("punch")) && (message_content.includes("kirimaru")) || message_content.includes("<@"+global_params.env_vars.OWNER_ID+">")){
 		//fetch messages sent by bot after this command
 		global_params.message.channel.fetchMessages({limit:5}).then(messages => {
         	const botMessages = messages.filter(msg => msg.author.bot);
-        	global_params.logger.debug(botMessages);
+        	/*global_params.logger.debug(botMessages);*/
         	global_params.message.channel.bulkDelete(botMessages).then(function(messages){
+        		global_params.message.channel.send("No one is allowed to slap or punch my mastah!!").then(function(){
+	        		global_params.logger.debug("Reply success");	
+	        	}).catch(function(){
+	        		global_params.logger.debug("Failed to send message an error occured");	
+	        	});
         		global_params.message.delete().then(function(message){
-	        		global_params.message.channel.send("No one is allowed to slap or punch my mastah!!").then(function(){
-	        			global_params.logger.debug("Reply success");	
-	        		}).catch(function(){
-	        			global_params.logger.debug("Failed to send message an error occured");	
-	        		});
 	        	}).catch(function(error){
 	        		global_params.logger.debug(error);		
 	        	});   
@@ -32,7 +32,7 @@ module.exports = (global_params,global_settings,message_type) => {
     	}).catch(function(error){
     		global_params.logger.debug(error);
     	});
-	}else */if(global_params.message.author.id === "232221425329504257" && global_settings.f2p === true){
+	}else if(global_params.message.author.id === "232221425329504257" && global_settings.f2p === true){
 		/*emi case 232221425329504257 */
 		try{
 			global_params.logger.debug("user matched");
